refactor: modify structure
