--- ewm-main-service/src/main/java/ru/practicum/ExploreWithMeMainService.java
@@ -1,11 +1,0 @@
-package ru.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ExploreWithMeMainService {
-    public static void main(String[] args) {
-        SpringApplication.run(ExploreWithMeMainService.class, args);
-    }
-}
