--- ewm-main-service/src/main/java/ru/practicum/category/model/Category.java
@@ -1,17 +1,0 @@
-package ru.practicum.category.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "categories")
-@AllArgsConstructor
-@NoArgsConstructor
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String name;
-}
