--- ewm-main-service/src/main/java/ru/practicum/category/repository/CategoryRepository.java
@@ -1,10 +1,0 @@
-package ru.practicum.category.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.category.model.Category;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-
-}
