--- ewm-main-service/src/main/java/ru/practicum/enums/State.java
@@ -1,8 +1,0 @@
-package ru.practicum.enums;
-
-public enum State {
-    PENDING, // в ожидании
-    PUBLISHED, // опубликовано
-    CANCELED, // отменено
-    CONFIRMED, // подтверждено
-}
