--- ewm-main-service/src/main/java/ru/practicum/event/model/Location.java
@@ -1,18 +1,0 @@
-package ru.practicum.event.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "locations")
-@AllArgsConstructor
-@NoArgsConstructor
-public class Location {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private Float lat;
-    private Float lon;
-}
