+++ ewm-main-service/src/main/java/ru/practicum/ewm/category/Category.java
@@ -1,0 +1,17 @@
+package ru.practicum.ewm.category;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "categories")
+@AllArgsConstructor
+@NoArgsConstructor
+public class Category {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+    private String name;
+}
