+++ ewm-main-service/src/main/java/ru/practicum/ewm/category/CategoryRepository.java
@@ -1,0 +1,9 @@
+package ru.practicum.ewm.category;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
+}
