+++ ewm-main-service/src/main/java/ru/practicum/ewm/exception/AccessException.java
@@ -1,0 +1,7 @@
+package ru.practicum.ewm.exception;
+
+public class AccessException extends RuntimeException {
+    public AccessException(String message) {
+        super(message);
+    }
+}
