+++ ewm-main-service/src/main/java/ru/practicum/ewm/exception/DataViolationException.java
@@ -1,0 +1,7 @@
+package ru.practicum.ewm.exception;
+
+public class DataViolationException extends RuntimeException {
+    public DataViolationException(String message) {
+        super(message);
+    }
+}
