+++ ewm-main-service/src/main/java/ru/practicum/ewm/request/dto/ParticipationRequestDto.java
@@ -1,0 +1,16 @@
+package ru.practicum.ewm.request.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ParticipationRequestDto {
+    private Long id;
+    private String created;
+    private Long event;
+    private Long requester;
+    private String status;
+}
