+++ ewm-main-service/src/main/java/ru/practicum/ewm/request/model/ParticipationRequest.java
@@ -1,0 +1,30 @@
+package ru.practicum.ewm.request.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+import ru.practicum.enums.State;
+import ru.practicum.event.model.Event;
+import ru.practicum.ewm.user.model.User;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "participation_requests")
+@AllArgsConstructor
+@NoArgsConstructor
+public class ParticipationRequest {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private LocalDateTime created;
+    @ManyToOne
+    @JoinColumn(name = "event_id")
+    private Event event;
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User requester;
+    @Enumerated(EnumType.STRING)
+    private State status;
+}
