+++ ewm-main-service/src/main/java/ru/practicum/ewm/request/service/RequestService.java
@@ -1,0 +1,13 @@
+package ru.practicum.ewm.request.service;
+
+import ru.practicum.ewm.request.dto.ParticipationRequestDto;
+
+import java.util.List;
+
+public interface RequestService {
+    ParticipationRequestDto create(Long userId, Long eventId);
+
+    List<ParticipationRequestDto> get(Long userId);
+
+    ParticipationRequestDto cancel(Long userId, Long requestId);
+}
