+++ ewm-main-service/src/main/java/ru/practicum/ewm/user/dto/UserShortDto.java
@@ -1,4 +1,13 @@
 package ru.practicum.ewm.user.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class UserShortDto {
+    private Long id;
+    private String name;
 }
