+++ ewm-main-service/src/main/java/ru/practicum/ewm/user/mapper/UserMapper.java
@@ -1,0 +1,30 @@
+package ru.practicum.ewm.user.mapper;
+
+import org.springframework.stereotype.Component;
+import ru.practicum.ewm.user.dto.UserDto;
+import ru.practicum.ewm.user.dto.UserShortDto;
+import ru.practicum.ewm.user.model.User;
+
+@Component
+public class UserMapper {
+
+    public User toUser(UserDto user) {
+        return new User(
+                user.getId(),
+                user.getEmail(),
+                user.getName());
+    }
+
+    public UserDto toUserDto(User user) {
+        return new UserDto(
+                user.getId(),
+                user.getEmail(),
+                user.getName());
+    }
+
+    public UserShortDto toUserShortDto(User user) {
+        return new UserShortDto(
+                user.getId(),
+                user.getName());
+    }
+}
