+++ ewm-main-service/src/main/java/ru/practicum/ewm/user/model/User.java
@@ -1,0 +1,18 @@
+package ru.practicum.ewm.user.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "users")
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String email;
+    private String name;
+}
