+++ ewm-main-service/src/main/java/ru/practicum/ewm/user/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package ru.practicum.ewm.user.repository;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.practicum.ewm.user.model.User;
+
+import java.util.Collection;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Page<User> findAllByIdIn(Collection<Long> id, Pageable pageable);
+
+    boolean existsByName(String name);
+}
