+++ ewm-main-service/src/main/java/ru/practicum/ewm/user/service/UserService.java
@@ -1,0 +1,13 @@
+package ru.practicum.ewm.user.service;
+
+import ru.practicum.ewm.user.dto.UserDto;
+
+import java.util.List;
+
+public interface UserService {
+    UserDto create(UserDto user);
+
+    List<UserDto> get(Long[] ids, int from, int size);
+
+    void delete(Long userId);
+}
