--- ewm-main-service/src/main/java/ru/practicum/exception/AccessException.java
@@ -1,7 +1,0 @@
-package ru.practicum.exception;
-
-public class AccessException extends RuntimeException {
-    public AccessException(String message) {
-        super(message);
-    }
-}
