--- ewm-main-service/src/main/java/ru/practicum/exception/DataViolationException.java
@@ -1,7 +1,0 @@
-package ru.practicum.exception;
-
-public class DataViolationException extends RuntimeException {
-    public DataViolationException(String message) {
-        super(message);
-    }
-}
