--- ewm-main-service/src/main/java/ru/practicum/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package ru.practicum.exception;
-
-public class NotFoundException extends RuntimeException {
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
