--- ewm-main-service/src/main/java/ru/practicum/request/controller/RequestController.java
@@ -1,32 +1,0 @@
-package ru.practicum.request.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-import ru.practicum.request.dto.ParticipationRequestDto;
-import ru.practicum.request.service.RequestService;
-
-import java.util.List;
-
-import static org.springframework.http.HttpStatus.CREATED;
-
-@RestController
-@RequiredArgsConstructor
-public class RequestController {
-    private final RequestService requestService;
-
-    @PostMapping("/users/{userId}/requests")
-    @ResponseStatus(CREATED)
-    public ParticipationRequestDto create(@PathVariable Long userId, @RequestParam Long eventId) {
-        return requestService.create(userId, eventId);
-    }
-
-    @GetMapping("/users/{userId}/requests")
-    public List<ParticipationRequestDto> get(@PathVariable Long userId) {
-        return requestService.get(userId);
-    }
-
-    @PatchMapping("/users/{userId}/requests/{requestId}/cancel")
-    public ParticipationRequestDto cancel(@PathVariable Long userId, @PathVariable Long requestId) {
-        return requestService.cancel(userId, requestId);
-    }
-}
