--- ewm-main-service/src/main/java/ru/practicum/request/repository/RequestRepository.java
@@ -1,18 +1,0 @@
-package ru.practicum.request.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.practicum.event.model.Event;
-import ru.practicum.request.model.ParticipationRequest;
-import ru.practicum.user.model.User;
-
-import java.util.List;
-
-@Repository
-public interface RequestRepository extends JpaRepository<ParticipationRequest, Long> {
-    boolean existsByEventAndRequester(Event event, User requester);
-
-    List<ParticipationRequest> findAllByRequester(User requester);
-
-    List<ParticipationRequest> findByEventId(Long eventId);
-}
