--- ewm-main-service/src/main/java/ru/practicum/user/dto/UserDto.java
@@ -1,22 +1,0 @@
-package ru.practicum.user.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDto {
-    private Long id;
-    @Email
-    @NotBlank
-    @Length(min = 6, max = 254)
-    private String email;
-    @NotBlank
-    @Length(min = 2, max = 250)
-    private String name;
-}
