--- ewm-main-service/src/main/java/ru/practicum/user/dto/UserShortDto.java
@@ -1,13 +1,0 @@
-package ru.practicum.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserShortDto {
-    private Long id;
-    private String name;
-}
