--- ewm-main-service/src/main/java/ru/practicum/user/model/User.java
@@ -1,18 +1,0 @@
-package ru.practicum.user.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "users")
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String email;
-    private String name;
-}
